chore(look&feel,apollo): move ItemFile component to FileUpload folder
