+++ client/apollo/css/src/apollo.css
@@ -9,7 +9,6 @@
 @import "./Form/ItemLabel/ItemLabelApollo.css";
 @import "./Form/InputTextAtom/InputTextAtomApollo.css";
 @import "./Form/InputText/InputTextApollo.css";
-@import "./Form/ItemFile/ItemFileApollo.css";
 @import "./Form/TextArea/TextAreaApollo.css";
 @import "./Form/Dropdown/DropdownApollo.css";
 @import "./Icon/IconApollo.css";
@@ -46,3 +45,4 @@
 @import "./Layout/Footer/FooterApollo.css";
 @import "./Svg/Svg.css";
 @import "./Form/FileUpload/InputFile/InputFileApollo.css";
+@import "./Form/FileUpload/ItemFile/ItemFileApollo.css";
