+++ client/apollo/react/src/Form/FileUpload/ItemFile/ItemFileApollo.tsx
@@ -1,12 +1,13 @@
-import "@axa-fr/design-system-apollo-css/dist/Form/ItemFile/ItemFileApollo.css";
 import deleteIcon from "@material-symbols/svg-400/outlined/delete.svg";
 import visibilityIcon from "@material-symbols/svg-400/outlined/visibility-fill.svg";
-import { ClickIcon } from "../../ClickIcon/ClickIconApollo";
-import { Icon } from "../../Icon/IconApollo";
-import { Spinner } from "../../Spinner/SpinnerApollo";
-import { ItemMessage } from "../ItemMessage/ItemMessageApollo";
+import { ClickIcon } from "../../../ClickIcon/ClickIconApollo";
+import { Icon } from "../../../Icon/IconApollo";
+import { Spinner } from "../../../Spinner/SpinnerApollo";
+import { ItemMessage } from "../../ItemMessage/ItemMessageApollo";
 import { ItemFileCommon } from "./ItemFileCommon";
 import type { ItemFileProps } from "./types";
+
+import "@axa-fr/design-system-apollo-css/dist/Form/FileUpload/ItemFile/ItemFileApollo.css";
 
 export { itemFileVariants } from "./ItemFileCommon";
