+++ client/apollo/react/src/index.ts
@@ -38,7 +38,6 @@ export { type OptionType } from "./Form/InputPhone/InputPhone.types";
 export { InputPhone } from "./Form/InputPhone/InputPhoneApollo";
 export { InputText } from "./Form/InputText/InputTextApollo";
 export { InputTextAtom } from "./Form/InputTextAtom/InputTextAtomApollo";
-export { ItemFile, itemFileVariants } from "./Form/ItemFile/ItemFileApollo";
 export { ItemLabel } from "./Form/ItemLabel/ItemLabelApollo";
 export {
   ItemMessage,
@@ -107,3 +106,7 @@ export { Tag, tagVariants, type TagVariants } from "./Tag/TagApollo";
 export { TimelineVertical } from "./TimelineVertical/TimelineVerticalApollo";
 export { Toggle } from "./Toggle/ToggleApollo";
 export { InputFile } from "./Form/FileUpload/InputFile/InputFileApollo";
+export {
+  ItemFile,
+  itemFileVariants,
+} from "./Form/FileUpload/ItemFile/ItemFileApollo";
